svn merge -r 37200:37231 http://svn.civicrm.org/civicrm/trunk

+++ CRM/Contact/Form/DedupeRules.php
@@ -74,10 +74,10 @@ class CRM_Contact_Form_DedupeRules extends CRM_Admin_Form
             $this->_defaults['threshold']   = $rgDao->threshold;
             $this->_contactType             = $rgDao->contact_type;
             $this->_defaults['level']       = $rgDao->level;
-            $this->_defaults['name']        = $rgDao->name;
             $this->_defaults['title']       = $rgDao->title;
             $this->_defaults['is_default']  = $rgDao->is_default;
             $this->_defaults['is_reserved'] = $rgDao->is_reserved;
+            $this->assign( 'isReserved', $rgDao->is_reserved );
             $ruleDao = new CRM_Dedupe_DAO_Rule();
             $ruleDao->dedupe_rule_group_id = $this->_rgid;
             $ruleDao->find();
@@ -111,33 +111,46 @@ class CRM_Contact_Form_DedupeRules extends CRM_Admin_Form
         $foo = CRM_Core_DAO::getAttribute('CRM_Dedupe_DAO_Rule', 'title');
         
         $this->add('text', 'title', ts('Rule Name'), array( 'maxlength' => 255, 'class' => 'huge'), true );
-        $this->add('text', 'name', ts('Internal Name'), array( 'maxlength' => 64, 'class' => 'huge'), true );
-        $this->addRule( 'name', ts('A duplicate matching rule with this name already exists. Please select another name.'), 
-                        'objectExists', array( 'CRM_Dedupe_DAO_RuleGroup', $this->_rgid, 'name' ) );
+        $this->addRule( 'title', ts('A duplicate matching rule with this name already exists. Please select another name.'), 
+                        'objectExists', array( 'CRM_Dedupe_DAO_RuleGroup', $this->_rgid, 'title' ) );
         $levelType = array(
                            'Fuzzy'  => ts('Fuzzy'),
                            'Strict' => ts('Strict')
                            );
         $ruleLevel = $this->add('select', 'level', ts('Level'), $levelType);
-
-        $this->add('checkbox', 'is_reserved', ts('Reserved?') );
         
         $default = $this->add('checkbox', 'is_default', ts('Default?') );
         if ( CRM_Utils_Array::value( 'is_default', $this->_defaults ) ) {
             $default->freeze();
             $ruleLevel->freeze();
         }
-        
-        for ($count = 0; $count < self::RULES_COUNT; $count++) {
-            $this->add('select', "where_$count", ts('Field'), array(null => ts('- none -')) + $this->_fields);
-            $this->add('text', "length_$count", ts('Length'), array('class' => 'two', 'style' => 'text-align: right'));
-            $this->add('text', "weight_$count", ts('Weight'), array('class' => 'two', 'style' => 'text-align: right'));
+
+        $disabled = array(); 
+        $reserved = $this->add('checkbox', 'is_reserved', ts('Reserved?') );
+        if ( CRM_Utils_Array::value( 'is_reserved', $this->_defaults ) ) {
+            $reserved->freeze();
+            $ruleLevel->freeze();
+            $disabled = array( 'disabled' => true );
         }
-        $this->add('text', 'threshold', ts("Weight Threshold to Consider Contacts 'Matching':"), array('class' => 'two', 'style' => 'text-align: right'));
+
+        $attributes = array('class' => 'two');
+        if ( !empty( $disabled ) ) {
+            $attributes = array_merge( $attributes, $disabled );
+        }
+
+        for ($count = 0; $count < self::RULES_COUNT; $count++) {
+            $this->add('select', "where_$count", ts('Field'), 
+                        array(null => ts('- none -')) + $this->_fields, false, $disabled );
+            $this->add('text', "length_$count", ts('Length'), $attributes );
+            $this->add('text', "weight_$count", ts('Weight'), $attributes );
+        }
+        
+        $this->add('text', 'threshold', ts("Weight Threshold to Consider Contacts 'Matching':"), $attributes );
         $this->addButtons(array(
                                 array('type' => 'next',   'name' => ts('Save'), 'isDefault' => true),
                                 array('type' => 'cancel', 'name' => ts('Cancel')),
                                 ));
+        
         $this->assign('contact_type', $this->_contactType);
 
         $this->addFormRule( array( 'CRM_Contact_Form_DedupeRules', 'formRule' ), $this );
@@ -155,6 +168,10 @@ class CRM_Contact_Form_DedupeRules extends CRM_Admin_Form
     static function formRule( $fields, $files, $self ) 
     {
         $errors = array( );
+        if ( CRM_Utils_Array::value( 'is_reserved', $fields ) ) {
+            return true;
+        }
+
         $fieldSelected = false;
         for ($count = 0; $count < self::RULES_COUNT; $count++) {
             if ( CRM_Utils_Array::value( "where_$count", $fields ) ) {
@@ -184,6 +201,7 @@ class CRM_Contact_Form_DedupeRules extends CRM_Admin_Form
     public function postProcess() 
     {
         $values = $this->exportValues();
+        
         $isDefault = CRM_Utils_Array::value( 'is_default', $values, false );
         // reset defaults
         if ( $isDefault ) {
@@ -198,18 +216,32 @@ UPDATE civicrm_dedupe_rule_group
         }
 
         $rgDao            = new CRM_Dedupe_DAO_RuleGroup();
-        if ($this->_action & CRM_Core_Action::UPDATE ) {
-            $rgDao->id           = $this->_rgid;
+        if ( $this->_action & CRM_Core_Action::UPDATE ) {
+            $rgDao->id    = $this->_rgid;
         }
-        $rgDao->threshold    = $values['threshold'];
+        
         $rgDao->title        = $values['title'];
-        $rgDao->name         = $values['name'];
-        $rgDao->level        = $values['level'];
-        $rgDao->contact_type = $this->_contactType;
         $rgDao->is_reserved  = CRM_Utils_Array::value( 'is_reserved', $values, false );
         $rgDao->is_default   = $isDefault;
+        $rgDao->level        = $values['level'];
+        $rgDao->contact_type = $this->_contactType;
+        $rgDao->threshold    = $values['threshold'];
         $rgDao->save();
         
+        // make sure name is set only during insert
+        if ( $this->_action & CRM_Core_Action::ADD ) {
+            // generate name based on title
+            require_once 'CRM/Utils/String.php';
+            $rgDao->name = CRM_Utils_String::titleToVar( $values['title'] ) . "_{$rgDao->id}";
+            $rgDao->save();
+        }
+        
+        // lets skip updating of fields for reserved dedupe group
+        if ( $rgDao->is_reserved ) {
+            CRM_Core_Session::setStatus( ts('The rule \'%1\' has been saved.', array(1 => $rgDao->title) ) ); 
+            return;
+        }
+
         $ruleDao = new CRM_Dedupe_DAO_Rule();
         $ruleDao->dedupe_rule_group_id = $rgDao->id;
         $ruleDao->delete();
@@ -259,9 +291,7 @@ UPDATE civicrm_dedupe_rule_group
         
         require_once 'CRM/Core/BAO/PrevNextCache.php';
         CRM_Core_BAO_PrevNextCache::deleteItem( null, $cacheKey );
-                
+
+        CRM_Core_Session::setStatus( ts('The rule \'%1\' has been saved.', array(1 => $rgDao->title) ) ); 
     }
-    
 }
-
-
